billa/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "lxcor-billa"
version = "0.1.0"
description = "Credit-based billing for Django APIs — pluggable payment backends (Stripe, PayPal, stub), credit pack catalogue, and per-key balance tracking."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "lxcor", email = "i.alves.alex@gmail.com" }]
requires-python = ">=3.8"
keywords = ["django", "billing", "stripe", "paypal", "credits", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "django>=3.2",
    "djangorestframework>=3.14",
    "lxcor-reggi>=0.1.0",
    "lxcor-kotta>=0.1.0",
]

[project.optional-dependencies]
stripe = ["stripe>=5.0"]
paypal = ["paypalrestsdk>=1.13"]

[project.urls]
Homepage = "https://github.com/lxcor/api-suite/tree/main/billa"
Repository = "https://github.com/lxcor/api-suite"

[tool.setuptools.packages.find]
where = ["."]
include = ["billa*"]
